+++ probes/issueActivityByProjectMember/def.yml
@@ -12,7 +12,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-id: activityOnIssuesByCollaboratorsMembersOrOwnersInLast90Days
+id: issueActivityByProjectMember
 short: Checks that a collaborator, member or owner has participated in issues in the last 90 days.
 motivation: >
   A project which is not active might not be patched, have its dependencies patched, or be actively tested and used. However, a lack of active maintenance is not necessarily always a problem. Some software, especially smaller utility functions, does not normally need to be maintained. For example, a library that determines if an integer is even would not normally need maintenance unless an underlying implementation language definition changed. A lack of active maintenance should signal that potential users should investigate further to judge the situation.
